feature: user query, command 클래스 추가
