+++ src/main/java/com/github/thundermarket/thundermarket/controller/UserController.java
@@ -6,7 +6,8 @@ import com.github.thundermarket.thundermarket.aspect.SessionUserParam;
 import com.github.thundermarket.thundermarket.constant.SessionConst;
 import com.github.thundermarket.thundermarket.domain.SessionUser;
 import com.github.thundermarket.thundermarket.domain.User;
-import com.github.thundermarket.thundermarket.service.UserService;
+import com.github.thundermarket.thundermarket.service.UserCommandHandler;
+import com.github.thundermarket.thundermarket.service.UserQueryHandler;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,15 +16,17 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 public class UserController {
 
-    private final UserService userService;
+    private final UserQueryHandler userQueryHandler;
+    private final UserCommandHandler userCommandHandler;
 
-    public UserController(UserService userService) {
-        this.userService = userService;
+    public UserController(UserQueryHandler userQueryHandler, UserCommandHandler userCommandHandler) {
+        this.userQueryHandler = userQueryHandler;
+        this.userCommandHandler = userCommandHandler;
     }
 
     @GetMapping("/api/v1/users")
     public ResponseEntity<?> findAllUsers() {
-        return new ResponseEntity<>(userService.findAllUsers(), HttpStatus.OK);
+        return new ResponseEntity<>(userQueryHandler.findAllUsers(), HttpStatus.OK);
     }
 
     @Authenticated
@@ -32,13 +35,13 @@ public class UserController {
         if(!user.isEmailValid()) {
             return new ResponseEntity<>(Email.NOT_VALID_EMAIL_MESSAGE, HttpStatus.BAD_REQUEST);
         }
-        return new ResponseEntity<>(userService.join(user), HttpStatus.OK);
+        return new ResponseEntity<>(userCommandHandler.join(user), HttpStatus.OK);
     }
 
     @Authenticated
     @PostMapping("/api/v1/auth/login")
     public ResponseEntity<?> login(@RequestBody User user, HttpSession session) {
-        User savedUser = userService.checkCredential(user);
+        User savedUser = userQueryHandler.checkCredential(user);
         if(savedUser == null) {
             return new ResponseEntity<>("Invalid credentials", HttpStatus.UNAUTHORIZED);
         }
